refactor: 녹음 버튼 컴포넌트 및 기능 분리
